SpawnPoint to float

+++ src/Engine/Events/EventInterpreter.cpp
@@ -68,7 +68,7 @@ static bool checkSeason(Season season) {
  * @offset 0x448CF4
  */
 static void spawnMonsters(int16_t typeindex, int16_t level, int count,
-                          Vec3i pos, int group, unsigned int uUniqueName) {
+                          Vec3f pos, int group, unsigned int uUniqueName) {
     MapId mapId = pMapStats->GetMapInfo(pCurrentMapName);
     SpawnPoint pSpawnPoint;
 
@@ -317,7 +317,7 @@ int EventInterpreter::executeOneEvent(int step, bool isNpc) {
             break;
         case EVENT_SummonMonsters:
             spawnMonsters(ir.data.monster_descr.type, ir.data.monster_descr.level, ir.data.monster_descr.count,
-                          Vec3i(ir.data.monster_descr.x, ir.data.monster_descr.y, ir.data.monster_descr.z),
+                          Vec3f(ir.data.monster_descr.x, ir.data.monster_descr.y, ir.data.monster_descr.z),
                           ir.data.monster_descr.group, ir.data.monster_descr.name_id);
             break;
         case EVENT_CastSpell:
